+++ src/workflows/workflows.ts
@@ -33,7 +33,12 @@ import {
   completeEthWithdrawalWorkflow,
   prepareWithdrawalWorkflow,
 } from './withdrawal';
-import { createOrderWorkflow, cancelOrderWorkflow } from './orders';
+import {
+  createOrderWorkflow,
+  cancelOrderWorkflow,
+  cancelOrderWorkflowWithSigner,
+  createOrderWorkflowWithSigner,
+} from './orders';
 import { createTradeWorkflow, createTradeWorkflowWithSigner } from './trades';
 import {
   UnsignedMintRequest,
@@ -55,6 +60,7 @@ import {
   L2Signer,
 } from '../types';
 import { Registration__factory } from '../contracts';
+import { WalletConnection } from '../types/index';
 
 export class Workflows {
   private readonly depositsApi: DepositsApi;
@@ -278,6 +284,7 @@ export class Workflows {
     }
   }
 
+  /** @deprecated */
   public createOrder(
     signer: Signer,
     starkWallet: StarkWallet,
@@ -286,12 +293,21 @@ export class Workflows {
     return createOrderWorkflow(signer, starkWallet, request, this.ordersApi);
   }
 
+  public createOrderWithSigner(walletConnection: WalletConnection, request: GetSignableOrderRequest) {
+    return createOrderWorkflowWithSigner({ ...walletConnection, request, ordersApi: this.ordersApi });
+  }
+
+  /** @deprecated */
   public cancelOrder(
     signer: Signer,
     starkWallet: StarkWallet,
     request: GetSignableCancelOrderRequest,
   ) {
     return cancelOrderWorkflow(signer, starkWallet, request, this.ordersApi);
+  }
+
+  public cancelOrderWithSigner(walletConnection: WalletConnection, request: GetSignableCancelOrderRequest) {
+    return cancelOrderWorkflowWithSigner({ ...walletConnection, request, ordersApi: this.ordersApi });
   }
 
   /** @deprecated */